Moved the FSDOMMutator interface in the right package.
