+++ openhtmltopdf-core/src/main/java/com/openhtmltopdf/extend/FSDOMMutator.java
@@ -1,6 +1,4 @@
-package com.openhtmltopdf.outputdevice.helper;
-
-import java.io.IOException;
+package com.openhtmltopdf.extend;
 
 /**
  * Allows to modify the HTML document DOM after it has been parsed
